Started adding compression
Server now serves compressed app.js bundle as app.js.gz

You will also notice in ./server/index.js that I added the
ability to compress the index.html file and serve that. You
will see this in the commented out sections. Those changes
are not live sicne they seem to make the server slower.
Keeping comments so I can test if deployed server is slower
with this addition. Also, since the index.html, for the proxy
is much larger, adding that code to proxy might actually be benificial